break the RangeSlider smart and UI portion as suggested by Trey.
